docs: add missing docs for memory-tracker crate
