+++ util/memory-tracker/src/jemalloc.rs
@@ -1,7 +1,13 @@
 use ckb_logger::info;
 use std::{ffi, mem, ptr};
 
-/// TODO(doc): @yangby-cryptape
+/// Dumps the heap through Jemalloc's API.
+///
+/// This funcions works when the following conditions are satisfied:
+/// - the global allocator is [Jemallocator].
+/// - the profiling is enabled.
+///
+/// [Jemallocator]: https://docs.rs/jemallocator/*/jemallocator/index.html
 pub fn jemalloc_profiling_dump(filename: &str) -> Result<(), String> {
     let mut filename0 = format!("{}\0", filename);
     let opt_name = "prof.dump";
