+++ util/memory-tracker/src/lib.rs
@@ -1,4 +1,5 @@
-//! TODO(doc): @yangby-cryptape
+//! Track the memory usage of CKB.
+
 #[cfg(all(
     not(target_env = "msvc"),
     not(target_os = "macos"),
@@ -11,7 +12,9 @@ mod jemalloc;
     feature = "profiling"
 )))]
 mod jemalloc {
-    /// TODO(doc): @yangby-cryptape
+    /// A dummy function which is used when the Jemalloc profiling isn't supported.
+    ///
+    /// Jemalloc profiling is disabled in default, the feature `profiling` is used to enable it.
     pub fn jemalloc_profiling_dump(_: &str) -> Result<(), String> {
         Err("jemalloc profiling dump: unsupported".to_string())
     }
@@ -26,7 +29,7 @@ mod process {
     use crate::rocksdb::TrackRocksDBMemory;
     use ckb_logger::info;
 
-    /// TODO(doc): @yangby-cryptape
+    /// A dummy function which is used when tracking memory usage isn't supported.
     pub fn track_current_process<Tracker: 'static + TrackRocksDBMemory + Sync + Send>(
         _: u64,
         _: Option<sync::Arc<Tracker>>,
@@ -34,13 +37,13 @@ mod process {
         info!("track current process: unsupported");
     }
 }
-pub mod rocksdb;
-pub mod utils;
+mod rocksdb;
 
 pub use jemalloc::jemalloc_profiling_dump;
 pub use process::track_current_process;
+pub use rocksdb::TrackRocksDBMemory;
 
-/// TODO(doc): @yangby-cryptape
+/// Track the memory usage of the CKB process and Jemalloc.
 pub fn track_current_process_simple(interval: u64) {
     track_current_process::<rocksdb::DummyRocksDB>(interval, None);
 }
